Move more helper functions to templates module
