+++ qubesadmin/tools/qvm_template.py
@@ -20,7 +20,6 @@
 """Tool for managing VM templates."""
 
 import argparse
-import base64
 import collections
 import configparser
 import datetime
@@ -61,7 +60,9 @@ from qubesadmin.templates import (
     Template,
     TemplateState,
     VersionSelector,
+    append_keys,
     build_version_str,
+    get_keys_for_repos,
     get_managed_template_vm,
     is_managed_template,
     is_match_spec,
@@ -74,8 +75,6 @@ TEMP_DIR = '/var/tmp'
 CACHE_DIR = os.path.join(xdg.BaseDirectory.xdg_cache_home, 'qvm-template')
 UNVERIFIED_SUFFIX = '.unverified'
 LOCK_FILE = '/run/qubes/qvm-template.lock'
-WRAPPER_PAYLOAD_BEGIN = "###!Q!BEGIN-QUBES-WRAPPER!Q!###"
-WRAPPER_PAYLOAD_END = "###!Q!END-QUBES-WRAPPER!Q!###"
 
 UPDATEVM = str('global UpdateVM')
 
@@ -305,53 +304,6 @@ def qrexec_popen(
         stderr=subprocess.PIPE
     )
 
-def _is_file_in_repo_templates_keys_dir(path: str) -> bool:
-    """Check if the given path is a file located repo-template keys dir"""
-    return os.path.isfile(path) and path.startswith(
-        "/etc/qubes/repo-templates/keys/")
-
-def _encode_key(path):
-    """Base64-encoe a file to be placed in qvm-template payload"""
-    if path.startswith("file://"):
-        path = path[7:]
-
-    if not _is_file_in_repo_templates_keys_dir(path):
-        return ""
-
-    encoded_key = "#" + path + "\n"
-    with open(path, "rb") as key:
-        encoded_key += f"#{base64.b64encode(key.read()).decode('ascii')}\n"
-    return encoded_key
-
-def _replace_dnf_vars(path, releasever):
-    """Replace supported dnf variables in repo"""
-    for var in ["$releasever", "${releasever}"]:
-        path = path.replace(var, releasever)
-    return path
-
-def _append_keys(payload, releasever):
-    """Add GPG key and SSL cert/keys to qvm-template payload"""
-    config = configparser.ConfigParser()
-    try:
-        config.read_string(payload)
-    except RuntimeError:
-        return ""
-
-    file_list = set()
-    for section in config.sections():
-        for option in ["gpgkey", "sslclientcert", "sslclientkey"]:
-            if config.has_option(section, option):
-                file_list.add(
-                    _replace_dnf_vars(config.get(section, option),
-                                      releasever))
-
-    encoded_keys = "".join(
-        [_encode_key(file_path) for file_path in sorted(file_list)])
-    if not encoded_keys:
-        return ""
-
-    return f"\n{WRAPPER_PAYLOAD_BEGIN}\n{encoded_keys}{WRAPPER_PAYLOAD_END}"
-
 def qrexec_payload(args: argparse.Namespace, app: qubesadmin.app.QubesBase,
                    spec: str, refresh: bool) -> str:
     """Return payload string for the ``qubes.Template*`` qrexec calls.
@@ -395,7 +347,7 @@ def qrexec_payload(args: argparse.Namespace, app: qubesadmin.app.QubesBase,
             repo_config += fd.read() + '\n'
     payload += repo_config
 
-    payload += _append_keys(repo_config, args.releasever)
+    payload += append_keys(repo_config, args.releasever)
     return payload
 
 
@@ -558,28 +510,6 @@ def qrexec_download(
                     "qrexec call 'qubes.TemplateDownload' failed.")
             if rpmcanon.wait() != 0:
                 raise ConnectionError("rpmcanon failed")
-
-
-def get_keys_for_repos(repo_files: typing.List[str],
-                       releasever: str) -> typing.Dict[str, str]:
-    """List gpg keys
-
-    Returns a dict reponame -> key path
-    """
-    keys = {}
-    for repo_file in repo_files:
-        repo_config = configparser.ConfigParser()
-        repo_config.read(repo_file)
-        for repo in repo_config.sections():
-            try:
-                gpgkey_url = repo_config.get(repo, 'gpgkey')
-            except configparser.NoOptionError:
-                continue
-            gpgkey_url = gpgkey_url.replace('$releasever', releasever)
-            # support only file:// urls
-            if gpgkey_url.startswith('file://'):
-                keys[repo] = gpgkey_url[len('file://'):]
-    return keys
 
 
 def filter_version(
